// titmas-typescript-sdk/generated/src/models/PreflightResult.ts
export interface PreflightResult {
    /**
     *
     * @type {Date}
     * @memberof PreflightResult
     */
    completedAt: Date;
    /**
     *
     * @type {Array<ReasonCode>}
     * @memberof PreflightResult
     */
    reasonCodes: Array<ReasonCode>;
    /**
     *
     * @type {string}
     * @memberof PreflightResult
     */
    receiptId: string;
    /**
     *
     * @type {string}
     * @memberof PreflightResult
     */
    requestId: string;
    /**
     *
     * @type {Result}
     * @memberof PreflightResult
     */
    result: Result;
    /**
     *
     * @type {string}
     * @memberof PreflightResult
     */
    schemaId: string | null;
    /**
     *
     * @type {string}
     * @memberof PreflightResult
     */
    schemaVersion: string | null;
}